refactor(docs): usage dx
