+++ apps/docs/content/components/user/link-description.raw.jsx
@@ -1,0 +1,17 @@
+import {User, Link} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <User
+      avatarProps={{
+        src: "https://avatars.githubusercontent.com/u/30373425?v=4",
+      }}
+      description={
+        <Link isExternal href="https://x.com/jrgarciadev" size="sm">
+          @jrgarciadev
+        </Link>
+      }
+      name="Junior Garcia"
+    />
+  );
+}
