+++ apps/docs/content/components/user/link-description.ts
@@ -1,20 +1,4 @@
-const App = `import {User, Link} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <User
-      name="Junior Garcia"
-      description={(
-        <Link href="https://x.com/jrgarciadev" size="sm" isExternal>
-          @jrgarciadev
-        </Link>
-      )}
-      avatarProps={{
-        src: "https://avatars.githubusercontent.com/u/30373425?v=4"
-      }}
-    />
-  );
-}`;
+import App from "./link-description.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
