+++ apps/docs/content/components/user/usage.raw.jsx
@@ -1,0 +1,13 @@
+import {User} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <User
+      avatarProps={{
+        src: "https://i.pravatar.cc/150?u=a04258114e29026702d",
+      }}
+      description="Product Designer"
+      name="Jane Doe"
+    />
+  );
+}
