+++ apps/docs/content/components/user/usage.ts
@@ -1,16 +1,4 @@
-const App = `import {User} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <User   
-      name="Jane Doe"
-      description="Product Designer"
-      avatarProps={{
-        src: "https://i.pravatar.cc/150?u=a04258114e29026702d"
-      }}
-    />
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
